seperate out coordinate object
